Add weather refresh infrastructure to PKJS
This implementation relies on localStorage to track the time of the
latest weather update. Each minute (using an interval with callback) the
time is compared against the most recent refresh, and when it exceeds a
threshold it is updated. Also, if the timestamp hasn't been set the
weather is initially updated.